# backend/openapi.yml
openapi: "3.0.2"
info:
  title: Rest API Docs
  version: "1.0"
servers:
  - url: http://localhost:4444/
  - url: https://ciaro360-server.vercel.app/
components:
  schemas:
    ResponseWithMsg:
      type: object
      properties:
        status:
          type: boolean
        msg:
          type: string
...
    ResponseWithMsgError:
      type: object
      properties:
        status:
          type: boolean
          example: false
        msg:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /control/create:
    post:
      tags:
        - "Control Routes"
      description: "Create a control"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                name:
                  type: string
                group:
                  type: string
      responses:
        "200":
          description: Control created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        "404":
          description: Control not created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized
      security:
        - bearerAuth: []
  /control/{id}:
    put:
      tags:
        - "Control Routes"
      description: "Update a control"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the control to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      responses:
        "200":
          description: Control updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        "404":
          description: Control not found or not updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
        401:
          description: Unauthorized
      security:
        - bearerAuth: []

    delete:
      tags:
        - "Control Routes"
      description: "Delete a control"
      parameters:
        - in: path
          name: id
          required: true
          description: ID of the control to delete
          schema:
            type: string
      responses:
        "200":
          description: Control deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        "404":
          description: Control not found or not deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

        401:
          description: Unauthorized

      security:
        - bearerAuth: []
    
    post:
      summary: Update a control
      tags:
        - "Control Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects associated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or risks association failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /control/groups:
    get:
      summary: Get all control groups
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          schema:
            type: integer
        - name: count
          in: query
          description: Number of items per page
          required: true
          schema:
            type: integer
        - name: search
          in: query
          description: Search query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
        '404':
          description: Not Found
  /control/subControls:
    get:
      summary: Get sub-controls
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: nameId
          required: true
          schema:
            type: string
        - in: query
          name: groupId
          required: true
          schema:
            type: string

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  controls:
                    type: array
                    items:
                      type: object
                      properties:
                        nameId:
                          type: string
                        group:
                          type: string
                        name:
                          type: string
                        content:
                          type: string
                        policies:
                          type: string
                        projects:
                          type: string
                        risks:
                          type: string
                        tags:
                          type: string
                        processDocs:
                          type: string
                        evidences:
                          type: string
                        custom:
                          type: boolean
        '404':
          description: Not Found
  /control/:
    get:
      summary: Get all controls
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          schema:
            type: integer
        - name: count
          in: query
          description: Number of items per page
          required: true
          schema:
            type: integer
        - name: search
          in: query
          description: Search query
          required: false
          schema:
            type: string
        - name: group
          in: query
          description: Group ID
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  controls: 
                    type: array
                    items:
                      type: object
                      properties:
                        nameId:
                          type: string
                        group:
                          type: string
                        name:
                          type: string
                        content:
                          type: string
                        policies:
                          type: string
                        projects:
                          type: string
                        risks:
                          type: string
                        tags:
                          type: string
                        processDocs:
                          type: string
                        evidences:
                          type: string
                        custom:
                          type: boolean
                  
        '404':
          description: Not Found
  /control/{id}/procedures:
    post:
      summary: Associate procedures with a control
      security:
        - bearerAuth: []
      tags:
        - "Control Routes"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                procedures:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Procedures associated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Control not found or procedures association failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      summary: Unassign procedures with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                procedures:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Procedures Unassign successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Control not found or procedures Unassign failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /control/{id}/projects:
    post:
      summary: Associate projects with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects associated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or projects association failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      summary: Unassign projects with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projects:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects unassign successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or projects Unassign failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /control/{id}/policies:
    post:
      summary: Associate policies with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policies:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects associated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or policies association failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
    delete:
      summary: Unassign policies with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                policies:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects Unassign successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or policies Unassign failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
  /control/{id}/risks:
    post:
      summary: Associate risks with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                risks:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects associated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or risks association failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
    delete:
      summary: Unassign risks with a control
      tags:
        - "Control Routes"
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                risks:
                  type: array
                  items:
                    type: string
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the control
      responses:
        '200':
          description: Projects Unassign successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: Control not found or risks Unassign failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"


  # Tags Routes
...
  /policy/{id}/controls:
    post:
      tags:
        - "Policy Routes"
      summary: "Map one policy to many controls"
      description: "This endpoint maps one policy to multiple controls."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                  description: "Array of control IDs."
              required:
                - controls
      responses:
        '200':
          description: "Mapping successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"

        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      tags:
        - "Policy Routes"
      summary: "Unmap controls from policy"
      description: "This endpoint unmaps controls from a specific policy."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the policy."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                  description: "Array of control IDs to unmap."
      responses:
        '200':
          description: "Controls unmapped from policy successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
...
  /risk/{id}/controls:
    post:
      tags:
        - "Risk Routes"
      summary: "Map controls to risk"
      description: "This endpoint maps controls to a risk."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the risk."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                  description: "Array of control IDs to map."
      responses:
        '200':
          description: "Controls mapped to risk successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"

    delete:
      tags:
        - "Risk Routes"
      summary: "Unmap controls from risk"
      description: "This endpoint unmaps controls from a risk."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The ID of the risk."
      responses:
        '200':
          description: "Controls unmapped from risk successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: "Resource not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
...
  /evidences/{id}/controls:
    put:
      summary: Link an evidence to one or more controls
      tags:
        - Evidences
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: The ID of the evidence to link to controls.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                    description: The IDs of the controls to link to the evidence.
              required:
                - controls
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []

    delete:
      summary: Remove the evidence from one or more controls
      tags:
        - Evidences
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            description: The ID of the evidence to remove from controls.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
                    description: The IDs of the controls from which to remove the evidence.
              required:
                - controls
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseWithMsgError'
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
...
  /project/{id}/controls:
    post:
      summary: Map controls to a project
      tags:
        - Project Routes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
      security:
        - bearerAuth: []

    delete:
      summary: Unlink controls from a project
      tags:
        - Project Routes
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                controls:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsg"
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseWithMsgError"
      security:
        - bearerAuth: []
...